gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic-settings==2.5.0
redis[hiredis]==5.1.0
httpx==0.27.0
langgraph==0.2.0
python-jose[cryptography]==3.3.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.0
bcrypt==4.2.0
email-validator
python-multipart
duckduckgo-search
langchain-ollama